+++ CHANGELOG.md
@@ -11,8 +11,13 @@
 
 - Renamed `Button` variant from `primary` to `solid`. ([KRAD-170])
 
+### Removed
+
+- `textDark` color inside the `theme` object. ([KRAD-169])
+
 [krad-153]: https://akrade.atlassian.net/browse/KRAD-153
 [krad-170]: https://akrade.atlassian.net/browse/KRAD-170
+[krad-169]: https://akrade.atlassian.net/browse/KRAD-169
 
 ## [0.3.0]